+++ invenio_rdm_records/assets/semantic-ui/js/invenio_rdm_records/src/deposit/components/CommunitySelectionModal/index.js
@@ -1,5 +1,5 @@
 // This file is part of Invenio-RDM-Records
-// Copyright (C) 2020-2023 CERN.
+// Copyright (C) 2020-2024 CERN.
 //
 // Invenio-RDM-Records is free software; you can redistribute it and/or modify it
 // under the terms of the MIT License; see LICENSE file for more details.
@@ -8,5 +8,7 @@ export {
   CommunitySelectionModal,
   CommunitySelectionModalComponent,
 } from "./CommunitySelectionModal";
+
+export { CommunitySelectionSearch } from "./CommunitySelectionSearch";
 
 export { CommunityContext } from "./CommunityContext";