+++ invenio_rdm_records/resources/serializers/schemas.py
@@ -8,6 +8,7 @@
 """Base parsing functions for the various serializers."""
 
 from marshmallow import missing
+from pydash import py_
 
 
 class CommonFieldsMixin:
@@ -55,7 +56,8 @@ class CommonFieldsMixin:
 
     def get_titles(self, obj):
         """Get titles."""
-        return [obj["metadata"]["title"]]
+        title = py_.get(obj, "metadata.title")
+        return [title] if title else missing
 
     def get_identifiers(self, obj):
         """Get identifiers."""
@@ -67,7 +69,9 @@ class CommonFieldsMixin:
 
     def get_creators(self, obj):
         """Get creators."""
-        return [c["person_or_org"]["name"] for c in obj["metadata"].get("creators", [])]
+        return [
+            c["person_or_org"]["name"] for c in obj["metadata"].get("creators", [])
+        ] or missing
 
     def get_publishers(self, obj):
         """Get publishers."""